+++ components/collector/src/source_collectors/owasp_zap/security_warnings.py
@@ -14,10 +14,11 @@ class OWASPZAPSecurityWarnings(XMLFileSourceCollector):
     """Collector to get security warnings from OWASP ZAP."""
 
     async def _parse_entities(self, responses: SourceResponses) -> Entities:
-        """Override to parse the securty warnings from the XML."""
+        """Override to parse the security warnings from the XML."""
         entities: dict[str, Entity] = {}
         tag_re = re.compile(r"<[^>]*>")
         risks = cast(list[str], self._parameter("risks"))
+        count_alert_instances = cast(str, self._parameter("alerts")) == "alert instances"
         for alert in await self.__alerts(responses, risks):
             ids = [
                 alert.findtext(id_tag, default="") for id_tag in ("alert", "pluginid", "cweid", "wascid", "sourceid")
@@ -25,19 +26,18 @@ class OWASPZAPSecurityWarnings(XMLFileSourceCollector):
             name = alert.findtext("name", default="")
             description = tag_re.sub("", alert.findtext("desc", default=""))
             risk = alert.findtext("riskdesc", default="")
-            for alert_instance in alert.findall("./instances/instance"):
-                method = alert_instance.findtext("method", default="")
-                uri = self.__stable(hashless(URL(alert_instance.findtext("uri", default=""))))
-                key = md5_hash(f"{':'.join(ids)}:{method}:{uri}")
-                entities[key] = Entity(
-                    key=key,
-                    old_key=md5_hash(f"{':'.join(ids[1:])}:{method}:{uri}"),
-                    name=name,
-                    description=description,
-                    uri=uri,
-                    location=f"{method} {uri}",
-                    risk=risk,
-                )
+            entity_kwargs = dict(name=name, description=description, risk=risk)
+            if count_alert_instances:
+                for alert_instance in alert.findall("./instances/instance"):
+                    method = alert_instance.findtext("method", default="")
+                    uri = self.__stable(hashless(URL(alert_instance.findtext("uri", default=""))))
+                    key = md5_hash(f"{':'.join(ids)}:{method}:{uri}")
+                    old_key = md5_hash(f"{':'.join(ids[1:])}:{method}:{uri}")
+                    location = f"{method} {uri}"
+                    entities[key] = Entity(key=key, old_key=old_key, uri=uri, location=location, **entity_kwargs)
+            else:
+                key = md5_hash(f"{':'.join(ids)}")
+                entities[key] = Entity(key=key, **entity_kwargs)
         return list(entities.values())
 
     def __stable(self, url: URL) -> URL: